increase sanity
